After cleaning up the code and changing spatial resolution for pitch and adding macro generator flag
